Add the loop to play 5 rounds
Now, as the playRound function is finished, it is introduced
into the game function, which plays 5 rounds and keeps a score
for the winner of each round. The score is kept by another function.
When the game is finished, a conditional alerts the winner of the game
based on the total points.